chore: fix default.ts
